renamed blog and post files
