--- client/src/components/blog/Content.js
@@ -1,17 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-
-const ContentContainer = styled.div`
-    margin-top: 0.5rem; 
-    width: 90%;
-`
-
-function Content() {
-    return (
-        <ContentContainer>
-            <p>Lorem ipsum dolor, sit amet consectetur adipisicing elit. Aspernatur nemo quam eveniet sit quae vel vitae veritatis ratione mollitia molestiae, alias ipsam voluptate natus inventore? Repudiandae quas consequuntur veritatis in, doloremque mollitia voluptatem perferendis voluptas provident velit, excepturi nemo reprehenderit voluptatum animi fuga eveniet obcaecati at aliquam architecto unde odit dolorem quod qui. Magnam doloremque neque iste blanditiis ipsum rerum nemo repudiandae facilis! Porro, in quisquam totam ipsa aliquam officia fugiat placeat eaque tenetur nostrum minus tempora, itaque natus est sint soluta suscipit? Quae officia explicabo, inventore in laboriosam ipsum voluptatibus tempora labore aperiam molestias fugiat nulla mollitia velit nisi et, voluptatem dolore error odit eius vero nostrum, reiciendis vel eos quis. Iste, eum enim inventore quidem, ipsa ex at quam quod neque doloribus nihil atque ratione dolores quo non facilis numquam dolore itaque. Velit tempore modi eos doloremque commodi ullam corporis illum sit dolore repellat. Voluptatem enim nihil ullam.</p>
-        </ContentContainer>
-    )
-}
-
-export default Content
